+++ src/main/java/max/exception/MaxException.java
@@ -1,3 +1,5 @@
+package max.exception;
+
 public class MaxException extends Exception {
     public MaxException(String message) {
         super(message);
