+++ src/main/java/max/task/Task.java
@@ -1,3 +1,5 @@
+package max.task;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
