+++ src/main/java/max/task/TaskList.java
@@ -1,3 +1,4 @@
+package max.task;
 import java.util.ArrayList;
 import java.util.List;
