+++ src/main/java/max/task/ToDo.java
@@ -1,3 +1,5 @@
+package max.task;
+
 public class ToDo extends Task {
