+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello I'm Max!
+Hello I'm max.Max!
  It's so nice to meet you!!
 
 Got it. I've added this task:
